SS4.0: Added ACL to sermons backend view as a test
git-svn-id: http://joomlacode.org/svn/sermon_speaker@248 5d6a53e0-5f49-0410-96cc-992822d07570

+++ SermonSpeaker4.0/com_sermonspeaker/admin/views/sermons/view.html.php
@@ -35,11 +35,16 @@ class SermonspeakerViewSermons extends JView
 	 */
 	protected function addToolbar()
 	{
+		$canDo 	= SermonspeakerHelper::getActions();
 		$state	= $this->get('State');
 
 		JToolBarHelper::title(JText::_('COM_SERMONSPEAKER_SERMON_TITLE'), 'sermons');
-		JToolBarHelper::addNew('sermon.add','JTOOLBAR_NEW');
-		JToolBarHelper::editList('sermon.edit','JTOOLBAR_EDIT');
+		if ($canDo->get('core.create')) {
+			JToolBarHelper::addNew('sermon.add','JTOOLBAR_NEW');
+		}
+		if ($canDo->get('core.edit')) {
+			JToolBarHelper::editList('sermon.edit','JTOOLBAR_EDIT');
+		}
 		JToolBarHelper::divider();
 		JToolBarHelper::custom('sermons.publish', 'publish.png', 'publish_f2.png','JTOOLBAR_PUBLISH', true);
 		JToolBarHelper::custom('sermons.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
@@ -52,12 +57,16 @@ class SermonspeakerViewSermons extends JView
 				JToolBarHelper::unarchiveList('sermons.publish', 'JTOOLBAR_UNARCHIVE');
 			}
 		}
-		if ($state->get('filter.state') == -2) {
-			JToolBarHelper::deleteList('', 'sermons.delete','JTOOLBAR_EMPTY_TRASH');
-		} else {
-			JToolBarHelper::trash('sermons.trash','JTOOLBAR_TRASH');
+		if ($canDo->get('core.delete')) {
+			if ($state->get('filter.state') == -2) {
+				JToolBarHelper::deleteList('', 'sermons.delete','JTOOLBAR_EMPTY_TRASH');
+			} else {
+				JToolBarHelper::trash('sermons.trash','JTOOLBAR_TRASH');
+			}
 		}
-		JToolBarHelper::divider();
-		JToolBarHelper::preferences('com_sermonspeaker');
+		if ($canDo->get('core.admin')) {
+			JToolBarHelper::divider();
+			JToolBarHelper::preferences('com_sermonspeaker');
+		}
 	}
 }
